Refactor events setup to offline class
